[UI] Add collapsing toolbar layout
The collapsing toolbar layout was going to be used for the image view for
each individual program. It is not working the way the docs describe or
perhaps I am not using it properly. More research into this view ir
required if I am to use it in an app. For now I am committing to this
branch for further investigation while I finish the app with a different
layout.