[Foundation] Implement single ProgramView
This view will be launched any time a program is clicked. This includes
from the schedule activity, the media player bar, the programs activity,
and the individual host activity. This commit adds a collapsing toolbar
layout and a palette implementation with picasso that dynamically colors
the second toolbar based on the image. closes #41, closes #42