[BugFix] Use onHandleIntent for programView
The program view was not updating with new information when a different
program was clicked. This commit adds the onHandleIntent method. There
is also a workaround for the collapsing toolbar title that I posted. The
workaround will be used until the fix is released. Currently it is
scheduled for future release.
https://code.google.com/p/android/issues/detail?can=1&start=0&num=100&q=collapsing&colspec=ID%20Type%20Status%20Owner%20Summary%20Stars&groupby=&sort=&id=175808closes
#45