[Refactor] Make ProgramListAdapter more generic.
The ProgramListAdapter was previously used only to show the list of
programs for a given day. This adapter can be reused for the programs
activity. This commit removes the need for a clicklistener. Putting the
click listener in the viewholder for the program keep me from
duplicating click listener code.