[BugFix] crash when loading program image
This commit adds a null check to the program image before we attempt to
load it via picasso. There is also a small fix removing the progress bar
when there is a failed api call.. Comments aded to clarify a hacky
solution avoiding a bug in the collapsing toolbar layout. The design
library  issue is marked for future release.